test(ui): split defaults tab coverage
